+++ agrest-engine/src/test/java/io/agrest/meta/DefaultEntityTest.java
@@ -83,7 +83,9 @@ public class DefaultEntityTest {
         assertEquals("oa1", oe1.getAttributes().stream().map(AgAttribute::getName).collect(Collectors.joining(",")));
 
         assertEquals(2, oe1.getSubEntities().size());
-        assertTrue(oe1.getSubEntities().containsAll(List.of(e11, e12)));
+        assertEquals("o11,o12", oe1.getSubEntities().stream().map(AgEntity::getName).sorted().collect(Collectors.joining(",")));
+        assertFalse(oe1.getSubEntities().contains(e11), "Child entity must have been overlaid with parent changes");
+        assertFalse(oe1.getSubEntities().contains(e12), "Child entity must have been overlaid with parent changes");
     }
 
     @Test